backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
dj-database-url==3.1.2
